cgraph: rewrite agxbuf macros as functions
This provides stronger type safety with no loss of performance. Note that this
also alters the implied type signature of `agxbuflen` to have a more appropriate
return type. This allows dropping one cast at a call site squashing another
warning.

+++ cmd/tools/gxl2gv.c
@@ -622,7 +622,7 @@ static void endElementHandler(void *userData, const char *name)
 
 	ud->closedElementType = TAG_NONE;
 	if (ud->compositeReadState) {
-	    size_t len = sizeof(GXL_COMP) + (size_t)agxblen(&ud->xml_attr_name);
+	    size_t len = sizeof(GXL_COMP) + agxblen(&ud->xml_attr_name);
 	    if (len <= SMALLBUF) {
 		new_name = buf;
 	    } else {
